% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Alon1999}
\alias{Alon1999}
\title{Gene expression data from Alon et al. (1999)}
\format{
A \code{data.frame} with 62 rows and 2001 variables:
\itemize{
\item \code{Y}: Grouping variable indicating whether a sample is from tumor
(\code{t}) or normal (\code{n}) tissue
\item \code{X1 - X2000}: Gene expression intensity levels for 2000 genes
}
}
\usage{
Alon1999
}
\description{
Gene expression levels from 40 tumor and 22 normal colon tissue samples. A
total of 6500 human genes were measured using an Affymetrix oligonucleotide
array. This dataset contains the 2000 genes with highest minimal intensity
across all samples. Note that the names of the genes have been omitted.
}
\references{
Alon, U. \emph{et al.} (1999). "Broad Patterns of Gene Expression Revealed by
Clustering Analysis of Tumor and Normal Colon Tissues Probed by
Oligonucleotide Arrays", PNAS \strong{96}(12), pp. 6745–50.
\href{https://doi.org/10.1073/pnas.96.12.6745}{doi:10.1073/pnas.96.12.6745}
}
\keyword{datasets}
