% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cheese}
\alias{cheese}
\title{Cheddar cheese taste}
\format{
A \code{data.frame} with 30 rows and 5 variables:
\describe{
\item{Case}{Sample number}
\item{Taste}{Subjective taste test score, obtained by combining the scores of several tasters}
\item{Acetic}{Natural log of concentration of acetic acid}
\item{H2S}{Natural log of concentration of hydrogen sulfide}
\item{Lactic}{Concentration of lactic acid}
}
}
\source{
\url{https://www3.nd.edu/~busiforc/handouts/Data\%20and\%20Stories/multicollinearity/Cheese\%20Taste/Cheddar\%20Cheese\%20Data.html}
}
\usage{
cheese
}
\description{
This dataset contains concentrations of various chemicals in 30 samples of
mature cheddar cheese, and a subjective measure of taste for each sample. The
variables \code{"Acetic"} and \code{"H2S"} are the natural logarithm of the
concentration of acetic acid and hydrogen sulfide respectively. The variable
\code{"Lactic"}, the concentration of lactic acid, has not been transformed.
}
\keyword{datasets}
