% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Einecke2010Kidney}
\alias{Einecke2010Kidney}
\title{Kidney transplant gene expression data from Einecke et al. (2010)}
\format{
A \code{data.frame} with patients (250) as rows and 10.001 features as columns:
\itemize{
\item \code{Reject_Status}: Binary variable indicating whether the transplanted
kidney was rejected (\code{1}) or not (\code{0}).
\item Remaining: gene expression levels for 10.000 genes. These were randomly
selected from the 25\% most variable genes in the original data
}
}
\usage{
Einecke2010Kidney
}
\description{
Gene expression data from patients receiving kidney transplantation, measured
using microarrays. The original data consists of 54675 genes from 282
patients. This data set contains a random subset of 250 patients and 10.000
of the 25\% most variable genes. For each patient, the rejection status of the
kidney transplant is recorded as well as the expression levels of each gene.
}
\references{
Einecke, G. \emph{et al.} (2010) "A molecular classifier for predicting future graft
loss in late kidney transplant biopsies", \emph{The Journal of Clinical
Investigation}, \strong{120}(6), pp. 1862–1872.
\href{https://doi.org/10.1172/JCI41789}{doi:10.1172/JCI41789}.
}
\keyword{datasets}
