% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{heavymetals}
\alias{heavymetals}
\title{Heavy metals near the river Schelde}
\format{
A \code{data.frame} with 19 rows and 18 variables:
\describe{
\item{location}{Location number}
\item{Cd5}{Concentration of Cadmium in mg/kg at 5 cm depth}
\item{Cd20}{Concentration of Cadmium in mg/kg at 20 cm depth}
\item{Cr5}{Concentration of Chromium in mg/kg at 5 cm depth}
\item{Cr20}{Concentration of Chromium in mg/kg at 20 cm depth}
\item{Cu5}{Concentration of Copper in mg/kg at 5 cm depth}
\item{Cu20}{Concentration of Copper in mg/kg at 20 cm depth}
\item{Fe5}{Concentration of Iron in mg/kg at 5 cm depth}
\item{Fe20}{Concentration of Iron in mg/kg at 20 cm depth}
\item{Mn5}{Concentration of Manganese in mg/kg at 5 cm depth}
\item{Mn20}{Concentration of Manganese in mg/kg at 20 cm depth}
\item{Ni5}{Concentration of Nickel in mg/kg at 5 cm depth}
\item{Ni20}{Concentration of Nickel in mg/kg at 20 cm depth}
\item{Pb5}{Concentration of Lead in mg/kg at 5 cm depth}
\item{Pb20}{Concentration of Lead in mg/kg at 20 cm depth}
\item{Zn5}{Concentration of Zinc in mg/kg at 5 cm depth}
\item{Zn20}{Concentration of Zinc in mg/kg at 20 cm depth}
\item{river}{Vicinity to the river: \code{0} (far) or \code{1} (close)}
}
}
\usage{
heavymetals
}
\description{
Concentrations of 8 heavy metals were measured on 19 different locations
along the river Schelde, each time at a depth of 5 cm and at a depth of 20
cm. Vicinity to the river was 0 (far) or 1 (close).
}
\keyword{datasets}
