% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{industries}
\alias{industries}
\title{Cold-war era employment in 26 European countries across different sectors}
\format{
A \code{data.frame} with 26 observations and 10 variables:
\describe{
\item{country}{The country name}
\item{agriculture}{Employment in the agriculture sector, in percentage}
\item{mining}{Employment in the mining sector, in percentage}
\item{manufacturing}{Employment in the manufacturing sector, in percentage}
\item{power.supply}{Employment in the power-supply sector, in percentage}
\item{construction}{Employment in the construction sector, in percentage}
\item{services}{Employment in the services sector, in percentage}
\item{finance}{Employment in the finance sector, in percentage}
\item{social.sector}{Employment in the social sector sector, in percentage}
\item{transport}{Employment in the transport sector, in percentage}
}
}
\usage{
data(industries)
}
\description{
Distribution of employment across 9 industrial sectors in 26 European
countries. The dataset stems from the Cold-War era. The values are the
percentages of the working population active in a given sector. Each row
should sum to 100.
}
\keyword{datasets}
