% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ukFoods}
\alias{ukFoods}
\title{Food consumption in the UK}
\format{
A \code{data.frame} with 17 rows representing the foodstuffs and 4 columns
representing each country:
\describe{
\item{England}{Average consumption in grams per person per week in England}
\item{Wales}{Average consumption in grams per person per week in Wales}
\item{Scotland}{Average consumption in grams per person per week in Scotland}
\item{N.Ireland}{Average consumption in grams per person per week in Northern Ireland}
}
}
\usage{
data(ukFoods)
}
\description{
This data set originates from the UKs ‘Department for Environment, Food and
Rural Affairs’ (DEFRA), showing the consumption in grams (per person, per
week) of 17 different types of foodstuff measured and averaged in the four
countries of the United Kingdom in 1997.
}
\keyword{datasets}
